IL-74 fix: Moved LocalStorageUtil file
